+++ src/oci-core/src/ui/commands/register-commands.ts
@@ -1,5 +1,5 @@
 /**
- * Copyright (c) 2022, 2023, Oracle and/or its affiliates. All rights reserved.
+ * Copyright © 2022, 2023, Oracle and/or its affiliates.
  * This software is licensed to you under the Universal Permissive License (UPL) 1.0 as shown at https://oss.oracle.com/licenses/upl.
  */
 
@@ -61,6 +61,8 @@ export async function registerNavigationCommands(context: ExtensionContext) {
             ext.treeDataProvider.refresh(undefined);
         }),
     );
+
+    commands.executeCommand('setContext', 'enableOCICoreViewTitleMenus', true);
 }
 
 export async function registerCommands(context: ExtensionContext, regionStatusBarItem: StatusBarItem, profileStatusBarItem: StatusBarItem) {