+++ src/oci-core/CHANGELOG.md
@@ -1,9 +1,19 @@
 # Change Log
 
-All notable changes to the "oci-core" extension will be documented in this file.
+All notable changes to the "faas" extension will be documented in this file.
 
 Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
 
 ## [1.0.0]
 
 - Initial release
+
+## 1.0.1 - 2023-06-22
+
+### Changed
+
+- Changed display name and description of the plugin.
+
+### Fixed
+
+- Enable view title menus only when tree view is loaded.