+++ src/orm-vscode-plugin/CHANGELOG.md
@@ -7,3 +7,20 @@ Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how 
 ## [1.0.0]
 
 - Initial release
+
+## 1.0.1 - 2023-06-22
+
+### Added
+
+- Added new notification banner when Resource Manager plugin is loaded.
+
+### Changed
+
+- Changed progress notification message when uploading stack changes to the RMS service.
+- Improved tooltip on stack selection in tree view.
+- Changed display name and description of the plugin.
+
+### Fixed
+
+- Enable view title menus only when tree view is loaded.
+